requirements.txt
asgiref==3.8.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
Django==5.2.3
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.1
requests==2.32.4
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0